Handle case where API response is compressed
The StackExchange API explicitly states that responses to requests are
gzipped.  For some reason, they are not zipped when called locally on my
machine, but they are zipped when run from Travis CI.

I do not know if `url-retrieve-synchronously' is performing any magic.
When I curl the request, I receive a gzipped response (as expected).  A
proper fix for this would be to somehow advise `url-*' to request as
curl does.