Fix test cases re file persistence
Every time the test was getting run, it would create the file, but not
delete it.  This meant that tests were not idempotent and artifacts from
previous tests were persisting into the next ones.

Use `make-temp-file' to let-bind `stack-filter-cache-file' to a
temporary file.  It is guaranteed to be empty when the tests begin.