Add basic HTML/JavaScript to enable authentication
This should be accessible using `raw.github.com/...', but if not, I
will put it on my webserver.  For now, this seems like it's as good as
it's going to get.

When the page is accessed with an access token, this token is
displayed (styled as CSS class `uservalue', big, dark, and green).

Error conditions are not currently being handled, but they will be.