Abstract out method calls
Keeping method calls within `sx-request.el' was causing circular
requirements.

This commit sorts through all of the requirements for each of the files
and ensures that this does not happen.  Much of the content removed was
for `sx-request-default-keyword-arguments-alist' and related items.  It
was unused, so it was pruned.  If it is deemed necessary in the future,
it should be included in `sx-method.el'.