Make site a url argument
This addresses the apparent api bug which affected both the retrieval of
favorites and the site-switching. Se #82

As far as I can see, this issue affects ALL api calls, so it seems
reasonable that it should be fixed inside `sx-method-call`.

The usage of sx-method-call stays the same, only the internal has changed.