Refactor .site to .site_par
It turns out some api objects do have a site property, except its value
is not a string, it is another object. The actual string we've been
referring to as .site is the .site.api_site_parameter

To avoid conflicts, I've renamed all our uses of .site to .site_par, and
sx-assoc-let now makes sure the object has a .site_par value, instead of
a .site value (which it may or may not have now, and is the same object
that the api refers to as site).