Don't rely on the new feature.
This is to avoid a bug in package.el which only gets fixed in Emacs 25
Unfortunately, there's no safe way to use a new feature offered by a
previously defined macro, because files were not reloaded on upgrade.

After a couple of weeks, once we can expect most of our users to have
restarted Emacs at least once since the last upgrade, we can start
relying on this new feature.